fix(fixtures): update fixtures to latest Zero usage
