+++ test/fixtures/_shared/.env
@@ -4,9 +4,9 @@ ZERO_LOG_LEVEL="debug"
 ZERO_ADMIN_PASSWORD="password"
 
 ZERO_MUTATE_FORWARD_COOKIES="true"
-ZERO_MUTATE_URL="http://localhost:3000/api/mutate"
+ZERO_MUTATE_URL="http://localhost:3000/api/zero/mutate"
 
 ZERO_QUERY_FORWARD_COOKIES="true"
-ZERO_QUERY_URL="http://localhost:3000/api/query"
+ZERO_QUERY_URL="http://localhost:3000/api/zero/query"
 
 ZERO_REPLICA_FILE="/tmp/zstart_replicas.db"
