+++ test/fixtures/nuxt/.env
@@ -1,2 +1,5 @@
-# NUXT_AUTH_SECRET can be set to keep login cookies stable across dev server restarts.
+NUXT_AUTH_SECRET="secretKey"
 NUXT_PUBLIC_ZERO_CACHE_URL="http://localhost:4848"
+NUXT_PUBLIC_ZERO_QUERY_URL="http://localhost:3000/api/zero/query"
+NUXT_PUBLIC_ZERO_MUTATE_URL="http://localhost:3000/api/zero/mutate"
+ZERO_UPSTREAM_DB="postgresql://user:password@127.0.0.1:5430/zstart"
