+++ test/fixtures/nuxt/app/app.vue
@@ -3,7 +3,7 @@ import { useInterval } from '#fx/composables/use-interval'
 import { randomMessage } from '#fx/db/data/test-data'
 import { formatDate } from '#fx/utils/date'
 import { randInt } from '#fx/utils/rand'
-import { mutators, queries } from '~/composables/zero'
+import { mutators, queries } from '#fx/db/schema'
 
 const zero = useZero()
 const { data: users } = useQuery(() => queries.users.all())
