+++ test/fixtures/nuxt/nuxt.config.ts
@@ -17,6 +17,8 @@ export default defineNuxtConfig({
     public: {
       zero: {
         cacheURL: '',
+        queryURL: '',
+        mutateURL: '',
       },
     },
   },
@@ -26,7 +28,12 @@ export default defineNuxtConfig({
       esbuildOptions: {
         target: 'es2022',
       },
-      include: ['@rocicorp/zero'],
+      include: [
+        '@rocicorp/zero',
+        '@rocicorp/zero/bindings',
+        'jose',
+        'zod',
+      ],
     },
   },
 })
