+++ test/fixtures/nuxt/package.json
@@ -18,7 +18,8 @@
   },
   "dependencies": {
     "jose": "^6.2.3",
-    "nuxt": "^4.2.1",
+    "nuxt": "^4.4.8",
+    "postgres": "3.4.7",
     "vue": "^3.5.34",
     "zero-vue": "link:",
     "zod": "4.4.3"
@@ -26,9 +27,9 @@
   "devDependencies": {
     "@nuxt/test-utils": "^4.0.3",
     "@types/node": "24.11.0",
-    "happy-dom": "^20.0.0",
+    "happy-dom": "^20.10.6",
     "typescript": "6.0.3",
     "vitest": "^4.1.7",
-    "vue-tsc": "3.2.8"
+    "vue-tsc": "3.3.6"
   }
 }
