+++ test/fixtures/vue/.env
@@ -1,3 +1,5 @@
 VITE_PUBLIC_ZERO_CACHE_URL="http://localhost:4848"
-# VUE_AUTH_SECRET can be set to keep login cookies stable across dev server restarts.
+VITE_PUBLIC_ZERO_QUERY_URL="http://localhost:3000/api/zero/query"
+VITE_PUBLIC_ZERO_MUTATE_URL="http://localhost:3000/api/zero/mutate"
+VITE_AUTH_SECRET="authSecret"
 ZERO_UPSTREAM_DB="postgresql://user:password@127.0.0.1:5430/zstart"
