+++ test/fixtures/vue/package.json
@@ -17,7 +17,7 @@
     "@vueuse/integrations": "^14.3.0",
     "h3": "^1.15.11",
     "jose": "^6.2.3",
-    "postgres": "^3.4.7",
+    "postgres": "3.4.7",
     "universal-cookie": "^8.1.2",
     "vue": "^3.5.34",
     "zero-vue": "link:",
@@ -30,7 +30,7 @@
     "@vitejs/plugin-vue": "6.0.7",
     "@vue/tsconfig": "0.9.1",
     "typescript": "6.0.3",
-    "vite": "8.0.14",
-    "vue-tsc": "3.2.8"
+    "vite": "8.1.3",
+    "vue-tsc": "3.3.6"
   }
 }
